Run e2e workflows on the context of the base branch
To allow PRs originating from forks access to secrets that are required
by the e2e test suite, it's required that the workflow runs in the
context of the base branch. As this poses a security risk (pwn
requests), the e2e tests will only run once the PR is approved so it is
ensured that it doesn't containe malintended code.